requirements/requirements.txt
timm==0.4.12
objprint==0.2.3
accelerate==0.23.0
torch==2.2.0
albumentations
tensorboard
easydict
SimpleITK
monai
tqdm
nibabel
pyyaml
openslide-python
opencv-python
openpyxl
pandas
histolab
pycox
lifelines
scikit-survival
scikit-learn
scanpy